# contract/openapi.yaml
openapi: 3.0.3
info:
  title: BuryMe API
  version: 1.0.0
  description: |
    Backend contract for **BuryMe** — a peer-to-peer IOU / micro-lending tracker.
    Derived directly from PRD v2.7, §5 (Conceptual Model), §6–7 (Requirements &
    Scenarios), §8 (Entity Attribute Reference), §11 (Dependencies & Integrations),
    and §12 (Validation & Error Handling).

    ## Architecture
    - **Thin clients.** All business logic, persistence, and third-party integration
      live in the Express/Node backend (§11.0). React (web) and Flutter (mobile)
      consume this API only — neither has direct DB or third-party access.
    - **Auth is Firebase-backed** (§11.1.2). Phone/email OTP and credential
      management happen entirely client-side via the Firebase Auth SDK. The client
      sends the resulting Firebase ID token as `Authorization: Bearer <token>` on
      every request; Express verifies it via the Firebase Admin SDK. There is
      therefore no `/login`, `/send-otp`, or `/verify-otp` endpoint here — those
      never touch this API.
    - **First login provisions a BuryMe user.** A valid Firebase token with no
      matching BuryMe profile yields `404` on `GET /auth/me`; the client then calls
      `POST /auth/register`.

    ## Request vs. Obligation — the core modeling decision
    Per §5 and §8.5/§8.6, **Request** and **Obligation** are two distinct persisted
    resources, not two states of one resource:
...
      It never re-enters a negotiation state.

...
    pre-filled.
...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /group-expenses:
    get:
      tags: [GroupExpenses]
      summary: List group expenses involving the current user
      operationId: listGroupExpenses
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [payer, participant] }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Page of group expenses.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/GroupExpense" }
                  next_cursor: { type: string, nullable: true }
    post:
      tags: [GroupExpenses]
      summary: Record a group expense and spawn per-participant obligations
      description: >
        Creates one bilateral Obligation per participant for their assigned share.
        No obligation is ever created for the payer's own share, even when
        `payer_share_included` is true (§8.12). Constraints: 1–50 participants,
        amounts must sum to the total within 0.01 ETB tolerance (§12.2.2).
        Each spawned obligation is `Lump Sum` / `Already Given` and `Active`
        from creation (the payer has already settled the bill), with
        `outstanding_balance` equal to that participant's assigned share and
        `due_date` taken from the input.
      operationId: createGroupExpense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGroupExpenseInput" }
      responses:
        "201":
          description: Group expense created with its participant obligations.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupExpense" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /group-expenses/{groupExpenseId}:
    get:
      tags: [GroupExpenses]
      summary: Get a group expense and its participants
      operationId: getGroupExpense
      parameters:
        - $ref: "#/components/parameters/GroupExpenseId"
      responses:
        "200":
          description: Group expense detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupExpense" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ══════════════════════════════════════════════════════════════
  # SETTLEMENTS (§6.5.2, §8.14)
  # ══════════════════════════════════════════════════════════════
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get the notification feed
      operationId: listNotifications
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Page of notifications, newest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Notification" }
                  next_cursor: { type: string, nullable: true }
  /notifications/read:
    post:
      tags: [Notifications]
      summary: Mark notifications as read
      operationId: markNotificationsRead
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notification_ids:
                  type: array
                  items: { type: string }
                  description: Omit to mark all as read.
      responses:
        "204": { description: Marked read. }
  /notifications/preferences:
    patch:
      tags: [Notifications]
      summary: Update notification preferences (§7.1.5)
      operationId: updateNotificationPreferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationPreferences" }
      responses:
        "200":
          description: Updated preferences.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Firebase ID token (1-hour expiry, §11.1.2). Verified on the backend via the
        Firebase Admin SDK on every protected route.
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
    GroupExpenseId:
      name: groupExpenseId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationError:
      description: One or more input fields failed validation (§12.3.2 `VALIDATION_ERROR`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_ERROR
              message: The due date must be a future date.
              field: proposed_due_date
...
    NotFound:
      description: Record does not exist or is not accessible to this user (§12.3.2 `NOT_FOUND`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    ErrorCode:
      type: string
      description: Standard error codes per §12.3.2.
      enum:
        - VALIDATION_ERROR
        - DUPLICATE_SUBMISSION
        - STATUS_CONFLICT
        - BALANCE_EXCEEDED
        - CONCURRENT_MODIFICATION
        - PAYMENT_GATEWAY_UNAVAILABLE
        - PAYMENT_GATEWAY_ERROR
        - RECIPIENT_UNVERIFIED
        - PARTIAL_WRITE_ROLLED_BACK
        - UNAUTHENTICATED
        - UNAUTHORIZED
        - NOT_FOUND
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Plain-language, user-safe. Never exposes stack traces, DB ids, or raw third-party payloads (§12.1).
            field:
              type: string
              nullable: true
              description: Present only for field-level validation errors; omitted for system errors (§12.3.1).

    # ── Enums (verbatim from §8) ─────────────────────────────────
...
    NotificationReferenceType:
      type: string
      description: What `reference_id` points at, so the client knows which screen to open.
      enum: [Request, Obligation, Payment, GroupExpense, SettlementSuggestion, TelebirrAccount]
    NotificationType:
      type: string
      description: >
        The §13.2 event catalog, keyed by its stable event id. The catalog is
        the full product spec; the ids marked "not emitted" below describe
        features that do not exist yet, and nothing raises them today:
        TB-03 (needs a `Pending Payout Setup` obligation status), RS-01/02/03
        (needs a schedule propose/accept/decline negotiation), OM-01/02/03
        (needs obligation change proposals), PAY-06 (needs Chapa reversal
        handling), GE-02 (needs obligation acknowledgement), BSS-01 (the
        platform never auto-detects a pair — suggestions are user-initiated).
        The four reminder events (RS-04, RS-05, RS-06, REM-01) are
        time-triggered per §13.3 and await a scheduler.
      enum:
        [
          TB-01,
          TB-02,
          TB-03,
          BR-01,
          BR-02,
          BR-03,
          BR-04,
          BR-05,
          BR-06,
          RR-01,
          GE-01,
          GE-02,
          RS-01,
          RS-02,
          RS-03,
          RS-04,
          RS-05,
          RS-06,
          OM-01,
          OM-02,
          OM-03,
          PAY-01,
          PAY-02,
          PAY-03,
          PAY-04,
          PAY-05,
          PAY-06,
          PAY-07,
          LR-01,
          LR-02,
          LR-03,
          LR-04,
          LR-05,
          LR-06,
          BSS-01,
          BSS-02,
          BSS-03,
          BSS-04,
          BSS-05,
          REM-01,
        ]

    # ── Shared value types ───────────────────────────────────────
    Money:
      type: object
      description: All amounts in ETB, ≤2 decimal places (§12.1).
      required: [amount, currency]
      properties:
        amount: { type: number, format: double, minimum: 0.01, maximum: 1000000, multipleOf: 0.01 }
        currency: { type: string, default: ETB, enum: [ETB] }

    # ── Auth / Users ──────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [user_id, display_name, profile_photo_url]
      properties:
        user_id: { type: string }
        display_name: { type: string }
        profile_photo_url: { type: string, nullable: true }
...
    CreateGroupExpenseInput:
      type: object
      description: §12.2.2 — 1–50 participants; assigned amounts (+ payer's own share, if included) must sum to total within 0.01 ETB.
      required:
        [total_amount, description, expense_date, due_date, payer_share_included, participants]
      properties:
        total_amount: { $ref: "#/components/schemas/Money" }
        description: { type: string, minLength: 3, maxLength: 200 }
        expense_date:
          type: string
          format: date
          description: Must not be in the future.
        due_date:
          type: string
          format: date
          description: >
            Repayment due date applied to every spawned obligation — on or after
            `expense_date`, at most 3 years out. The expense itself has no
            lifecycle; this is the deadline each participant owes by.
        payer_share_included: { type: boolean }
        payer_share_amount:
          allOf: [{ $ref: "#/components/schemas/Money" }]
          nullable: true
          description: Required when payer_share_included is true; omitted when false.
        participants:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [participant_user_id, assigned_amount]
            properties:
              participant_user_id: { type: string }
              assigned_amount: { $ref: "#/components/schemas/Money" }
      additionalProperties: false
    GroupExpenseParticipant:
      type: object
      required: [participant_entry_id, participant, assigned_amount, obligation_id]
      properties:
        participant_entry_id: { type: string }
        participant: { $ref: "#/components/schemas/UserSummary" }
        assigned_amount: { $ref: "#/components/schemas/Money" }
        obligation_id: { type: string }
    GroupExpense:
      type: object
      required:
        [
          expense_id,
          payer,
          total_amount,
          description,
          expense_date,
          due_date,
          payer_share_included,
          payer_share_amount,
          outstanding_total,
          participants,
          created_at,
        ]
      properties:
        expense_id: { type: string }
        payer: { $ref: "#/components/schemas/UserSummary" }
        total_amount: { $ref: "#/components/schemas/Money" }
        description: { type: string }
        expense_date: { type: string, format: date }
        due_date:
          type: string
          format: date
          description: Due date carried onto every spawned obligation.
        payer_share_included: { type: boolean }
        payer_share_amount:
          allOf: [{ $ref: "#/components/schemas/Money" }]
          nullable: true
          description: Recorded for accounting completeness only — no obligation is ever created against it.
        outstanding_total:
          allOf: [{ $ref: "#/components/schemas/Money" }]
          description: >
            Viewer-relative total still outstanding across this expense's
            spawned obligations: for the payer, the sum all participants
            still owe them; for a participant, their own remaining share.
            Excludes the payer's own share, which is never an obligation.
        participants:
          type: array
          items: { $ref: "#/components/schemas/GroupExpenseParticipant" }
        created_at: { type: string, format: date-time }

    # ── Settlement Suggestions (§8.14) ─────────────────────────────
...
    Notification:
      type: object
      required:
        [
          notification_id,
          notification_type,
          title,
          body,
          reference_id,
          reference_type,
          is_read,
          delivered_at,
          created_at,
        ]
      properties:
        notification_id: { type: string }
        notification_type: { $ref: "#/components/schemas/NotificationType" }
        title:
          type: string
          description: >
            Rendered at emit time from the §13.2 catalog's Title column.
            Stored rather than composed per client so the web app, the
            Flutter app and the FCM push payload all carry identical copy.
        body:
          type: string
          description: Rendered at emit time from the §13.2 catalog's Body column, with names, amounts and dates already interpolated.
        reference_id:
          type: string
          nullable: true
          description: The entity to open. Null for the catalog's Dashboard-targeted events (LR-06, BSS-05).
        reference_type:
          allOf: [{ $ref: "#/components/schemas/NotificationReferenceType" }]
          nullable: true
        is_read: { type: boolean }
        delivered_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    NotificationPreferences:
      type: object
      description: >
        Per-category push switches (§7.1.5, §6.1.5). Each maps to a family of
        §13.2 catalog events. These control push delivery only — every
        notification is still written to the in-app feed whatever is set
        here. All fields are optional on PATCH; omitted ones are left alone.
      required: [requests, group_expenses, payments, schedules, settlements, reminders, telebirr]
      properties:
        requests:
          type: boolean
          description: New borrow requests, counter-proposals, and lending records sent to you (BR-*, RR-01, LR-*).
        group_expenses:
          type: boolean
          description: Shares assigned to you, acknowledgements, and consolidated expense updates (GE-*).
        payments:
          type: boolean
          description: Chapa payments, external records waiting on your confirmation, and confirmed payments (PAY-*).
        schedules:
          type: boolean
          description: Schedules proposed to you, accepted schedules, and installment changes (RS-01 through RS-03).
        settlements:
          type: boolean
          description: Settlement suggestions, responses needed, and completed settlements (BSS-*).
        reminders:
          type: boolean
          description: Upcoming and overdue installments, and payments still waiting on a response (RS-04 through RS-06, REM-01).
        telebirr:
          type: boolean
          description: Prompts to verify Telebirr, and when someone you owe verifies theirs (TB-*).